--- src/utils/getImageUrl.js
@@ -1,1 +1,0 @@
-export const getImageUrl = (path) => `https://image.tmdb.org/t/p/w500${path}`;
